Added bbox field to each GeoJson type (#20)
The bounding box field is now extracted from the GeoJSON object if it exists. The API for interacting with GeoJSON also substantially changed forcing users to use the toplevel `to_json` and `of_json` which may return a feature, a collection of features or a geometry. This made parsing the shared fields (such as the bounding box field) much simpler.